Get transaction-based locks working
